vmci: djb2 in a signed int is undefined twice over
Chasing vmci_hashtable.c:170's `idx = VMCI_HASHTABLE_HASH(...)' to its
floor found something the index rule was not looking for.

The index is fine: vmci_hash_id() ends `return (hash & (size - 1))'
with size unsigned, so hash converts and the result is at most
size - 1 -- the same conversion that took 35 sites off the ONESIDED
list this morning, arrived at from the other direction.

The hash is not fine:

	int hash = 5381;
	for (i = 0; i < sizeof(id); i++)
		hash = ((hash << 5) + hash) + (uint8_t)(id >> (i * 8));

djb2 multiplies by 33 a round; from 5381 over four bytes it passes
INT_MAX on the third, and signed overflow is undefined. Once hash is
negative, `hash << 5' is undefined separately. This tree builds with
UBSan and its own smoke test checks that UBSan traps `1 << 31', so this
is the class the instruments exist for -- and the analyser says nothing
because it is not a path-sensitive property.

`unsigned int hash' fixes both. Unsigned wraparound is defined and
gives the identical bit pattern, so every caller's bucket is unchanged:
vmci_hashtable.c and vmci_doorbell.c's VMCI_DOORBELL_INDEX_TABLE_HASH
both.

Measured at HEAD: sys/dev/vmware analyser 2 -> 2 across 12 translation
units, all compiling both ways. No ONESIDED change -- the site that led
here still reports, because its floor is now a fact about a different
function, which is the shape this whole list is made of. Marker
revert-verified.

+++ hbsd/src/sys/dev/vmware/vmci/vmci_utils.h
@@ -30,12 +30,19 @@ static inline int
 vmci_hash_id(vmci_id id, unsigned size)
 {
 	unsigned i;
-	int hash = 5381;
+	/*
+	 * djb2 overflows within four rounds, and `hash' was a signed
+	 * int: both the overflow and the `hash << 5' that follows it on
+	 * a negative value are undefined.  Unsigned wraparound is
+	 * defined and produces the identical bit pattern, so the mask
+	 * below -- and every caller's bucket -- is unchanged.
+	 */
+	unsigned int hash = 5381;
 
 	for (i = 0; i < sizeof(id); i++)
 		hash = ((hash << 5) + hash) + (uint8_t)(id >> (i * 8));
 
-	return (hash & (size - 1));
+	return ((int)(hash & (size - 1)));
 }
 
 #endif /* !_VMCI_UTILS_H_ */
